WIP ### What's done: * WIP

+++ save-frontend/src/main/kotlin/com/saveourtool/save/frontend/components/views/OrganizationView.kt
@@ -601,7 +601,7 @@ class OrganizationView : AbstractView<OrganizationProps, OrganizationViewState>(
             topProject?.let {
                 scoreCard {
                     name = it.name
-                    contestScore = it.contestRating.toDouble()
+                    contestScore = it.contestRating
                     url = "#/${props.organizationName}/${it.name}"
                 }
             }
